requirements.txt
tensorflow==2.3.0
keras==2.3.1
numpy==1.18.5  
pillow==7.2.0
matplotlib==3.3.0
scikit-learn==0.23.2
opencv-python==4.4.0.42
boto3
flask
